% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_add_series.R
\name{c4a_add_series}
\alias{c4a_add_series}
\alias{c4a_add_series_as_is}
\alias{c4a_remove_series}
\title{Add and remove palette series}
\usage{
c4a_add_series(
  x,
  xNA = NA,
  types,
  series,
  format.palette.name = TRUE,
  remove.blacks = TRUE,
  take.gray.for.NA = TRUE,
  remove.other.grays = FALSE,
  light.to.dark = TRUE,
  remove.names = TRUE
)

c4a_add_series_as_is(
  ...,
  format.palette.name = FALSE,
  remove.blacks = FALSE,
  take.gray.for.NA = FALSE,
  remove.other.grays = FALSE,
  light.to.dark = FALSE,
  remove.names = FALSE
)

c4a_remove_series(series = "all", are.you.sure = FALSE)
}
\arguments{
\item{x}{named list of color palettes. See details for indexing.}

\item{xNA}{colors for missing values. Vector of the same length as x (or length 1). For `NA` values, the color for missing values is automatically determined.}

\item{types}{character vector of the same length as x (or length 1), which determines the type of palette `"cat"`, `"seq"`, or `"div"`.}

\item{series}{for `c4a_add_series`, a character vector of the same length as x (or length 1), which determines the series. For `c4a_remove_series` a character vector of series names that should be removed (use `"all"` to remove all).}

\item{format.palette.name}{should palette names be formatted to lowercase/underscore format?}

\item{remove.blacks, take.gray.for.NA, remove.other.grays}{These arguments determine the processing of grayscale colors for categorical `"cat"` palettes: if `remove.blacks` and there are (near) blacks, these are removed first. Next, if `take.gray.for.NA`, `xNA` is `NA`, and a palette contains at least one grayscale color (which can also be white), this is used as color for missing values. In case there are more than one grayscale color, the lightest is taken. `remove.other.grays` determines what happens with the other grays.}

\item{light.to.dark}{should sequential `"seq"` palettes be automatically ordered from light to dark?}

\item{remove.names}{should individual color names be removed?}

\item{...}{passed on to `c4a_add_series`}

\item{are.you.sure}{are you sure you want to remove series?}
}
\description{
Add and remove palette series. `c4a_add_series` can be used to add own palette series and  `c4a_remove_series` to remove palette series. `c4a_add_series_as_is` is the same as `c4a_add_series`, but by default without any processing.
}
\details{
Indexing: for a categorical `"cat"` palette, an optional `"index"` attribute determines which colors to use for which lengths: if the palette consists of k colors, index should be a list of k, where the i-th element is an integer vector of length i with values 1,2,...,k. See `cols4all::.z$palette$rainbow` and  for an example.
}
\examples{
cbsnl_cols = list(
	cool = c("#00a1cd", "#0058b8", "#afcb05", "#53a31d",
		"#f39200", "#af0e80", "#ffcc00", "#e94c0a"),
	warm = c("#e94c0a", "#ffcc00", "#af0e80", "#f39200",
		"#53a31d", "#afcb05", "#0058b8", "#00a1cd"),
	blues = c("#c0e7ff", "#77cbe5", "#3d95d4", "#2256a0", "#143564"),
	reds =  c("#ffc597", "#f89e6b", "#e74d15", "#c01f26", "#82001e"),
	greens = c("#edf0c7", "#c9de85", "#85bc22", "#348a3a", "#0f5f34"),
	purples = c("#f8c1d9", "#e38cbf", "#be3e8d", "#8b176f", "#460042")
)

cbsnl_types = c("cat", "cat", "seq", "seq", "seq", "seq")

c4a_add_series(cbsnl_cols, xNA = "grey88", types = cbsnl_types, series = "cbsnl")

\dontrun{
c4a_gui(series = "cbsnl")
}
}
