% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_convert_name.R
\name{c4a_convert_name}
\alias{c4a_convert_name}
\title{Converts a name to a known palette name, if it exists.}
\usage{
c4a_convert_name(name, no.match = c("error", "null"))
}
\arguments{
\item{name}{palette name to be converted}

\item{no.match}{what happens is no match is found? Options: `"error"`: an error is thrown, `"null"`: `NULL` is returned}
}
\description{
Converts a name to a known palette name, if it exists. Otherwise, either an error is thrown or `NULL` is returned
}
\examples{
c4a_convert_name("Blues")
c4a_convert_name("Jazz", no.match = "null")
c4a_convert_name("Rainbow")
c4a_convert_name("RdYlBu")
c4a_convert_name("magma")
}
