% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_gui.R, R/c4a_show.R
\name{c4a_gui}
\alias{c4a_gui}
\alias{c4a_show}
\title{Graphical user interface to analyse palettes}
\usage{
c4a_gui()

c4a_show(
  type = c("cat", "seq", "div"),
  n = NULL,
  cvd.sim = c("none", "deutan", "protan", "tritan"),
  sort = "name",
  text.col = "same",
  series = NULL,
  contrast = NULL,
  include.na = TRUE,
  show.scores = FALSE,
  columns = NA
)
}
\arguments{
\item{type}{type of palette: `"cat"` for categorical (aka qualitative), `"seq"` for sequential, and `"div"` for diverging}

\item{n}{number of colors. If omitted: for `"cat"` the full palette is displayed, and for `"seq"` and `"div"`, 9 colors.}

\item{cvd.sim}{color vision deficiency simulation: one of `"none"`, `"deutan"`, `"protan"`, `"tritan"`}

\item{sort}{column name to sort the data. For column names, see details}

\item{text.col}{The text color of the colors. By default `"same"`, which means that they are the same as the colors themselves (so invisible, but available for selection).}

\item{columns}{number of columns. By default equal to `n` or, if not specified, 12. Cannot be higher than the palette}
}
\description{
Graphical user interface to analyse palettes. `c4a_show` shows a table that can be opened in the browser. `c4a_gui` is a graphical user interface around this table.
}
\examples{
if (interactive()) {
	c4a_gui()
}

c4a_show(type = "cat")


# sort sequential palettes by rank (based on color blind-friendliness, and color balance)
c4a_show(type = "seq", n = 7, sort = "rank")

# sort by hue type, which is indication how many hues are used.
c4a_show(type = "seq", n = 5, sort = "hueType")
}
