% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_import_sysdata.R
\name{c4a_import_sysdata}
\alias{c4a_import_sysdata}
\alias{c4a_export_sysdata}
\title{Import and export system data}
\usage{
c4a_import_sysdata(x)

c4a_export_sysdata()
}
\arguments{
\item{x}{system data}
}
\value{
`c4a_export_sysdata` returns the system data
}
\description{
Import and export system data. `c4a_import_sysdata` will import system data and overwrite the current system data. System data includes palette data and pre-calculated color-blind-friendliness scores. `c4a_export_sysdata` will export the current system data. These functions are for advanced use; the functions `c4a_add_series` and `c4a_remove_series` are user-friendly interfaces to change the loaded palettes.
}
\examples{
x = c4a_export_sysdata()
c4a_import_sysdata(x)
y = c4a_export_sysdata()
identical(x, y)
}
