% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{c4a_ls}
\alias{c4a_ls}
\title{Nested list of cols4all palette names}
\format{
An object of class \code{environment} of length 11.
}
\usage{
c4a_ls
}
\description{
Nested list of cols4all palette names, which is handy for auto-completion
}
\keyword{datasets}
