% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_duplicate.R
\name{c4a_modify}
\alias{c4a_modify}
\alias{c4a_duplicate}
\title{Edit cols4all palettes}
\usage{
c4a_modify(palette, x = NULL, xNA = NULL)

c4a_duplicate(palette)
}
\arguments{
\item{palette}{name of the palette}

\item{x}{vector of the new colors. It should either the same length, or a named vector, where the names correspond to the index numbers. E.g. c("3" = "#AABBCC") will replace the third color with the color "#AABBCC".}

\item{xNA}{the new color for missing values.}
}
\description{
Edit cols4all palettes. c4a_duplicate duplicates an existing cols4all palette, and c4a_modify is used to change the colors. Use c4a_data to craete palettes from scratch.
}
\seealso{
\code{\link[=c4a_data]{c4a_data()}}
}
