% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_name_convert.R
\name{c4a_name_convert}
\alias{c4a_name_convert}
\title{Converts a name to a known palette name, if it exists.}
\usage{
c4a_name_convert(name, no.match = c("error", "null"), verbose = TRUE)
}
\arguments{
\item{name}{palette name to be converted}

\item{no.match}{what happens is no match is found? Options: \code{"error"}: an error is thrown, \code{"null"}: \code{NULL} is returned}

\item{verbose}{should messages be printed?}
}
\description{
Converts a name to a known palette name, if it exists. Otherwise, either an error is thrown or \code{NULL} is returned
}
\examples{
c4a_name_convert("Blues")
c4a_name_convert("Jazz", no.match = "null")
c4a_name_convert("Rainbow")
c4a_name_convert("RdYlBu")
c4a_name_convert("magma")
}
