% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_palettes.R
\docType{data}
\name{c4a_palettes}
\alias{c4a_palettes}
\alias{c4a_series}
\alias{c4a_defaults}
\title{List all available cols4all color palettes and series}
\format{
An object of class \code{character} of length 3.
}
\usage{
c4a_palettes(
  type = c("all", "cat", "seq", "div"),
  series = NULL,
  full.names = TRUE
)

c4a_series(type = c("all", "cat", "seq", "div"))

c4a_defaults
}
\arguments{
\item{type}{type of color palette: one of `"all"` (all palettes), `"cat"` (categorical/qualitative palettes), `"seq"` (sequential palettes) and `"div"` (diverging palettes).}

\item{series}{series to list the palettes from. Run `c4a_series` to see the options.}

\item{full.names}{should full names, i.e. with the prefix "series."? By default `TRUE`.}
}
\value{
names of the loaded color palettes
}
\description{
`c4a_palettes` lists all available cols4all color palettes. The used naming convention of the palettes is: series.palette_name. Palettes are also accessible (with \code{\link{c4a}})by just palette_name. `c4a_series` lists all series from which palettes are available. `c4a_defaults` returns the one default (recommended) palette per type
}
\examples{
c4a_palettes()

c4a_series()

c4a_palettes(type = "cat", series = "tol")

c4a_palettes(type = "seq", series = "kovesi")
}
\keyword{datasets}
