% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_palettes_add.R
\name{c4a_palettes_add}
\alias{c4a_palettes_add}
\alias{c4a_palettes_add_as_is}
\alias{c4a_palettes_remove}
\title{Add and remove palette series}
\usage{
c4a_palettes_add(
  x,
  xNA = NA,
  types,
  series,
  nmin = NA,
  nmax = NA,
  ndef = NA,
  mmin = NA,
  mmax = NA,
  mdef = NA,
  format.palette.name = TRUE,
  remove.blacks = TRUE,
  take.gray.for.NA = TRUE,
  remove.other.grays = FALSE,
  light.to.dark = TRUE,
  remove.names = TRUE,
  biv.method = "byrow",
  space = "rgb",
  range_matrix_args = list(NULL),
  bib = NA
)

c4a_palettes_add_as_is(
  ...,
  format.palette.name = FALSE,
  remove.blacks = FALSE,
  take.gray.for.NA = FALSE,
  remove.other.grays = FALSE,
  light.to.dark = FALSE,
  remove.names = FALSE
)

c4a_palettes_remove(fullnames = NULL, series = NULL, are.you.sure = NA)
}
\arguments{
\item{x}{named list of color palettes. See details for indexing.}

\item{xNA}{colors for missing values. Vector of the same length as x (or length 1). For \code{NA} values, the color for missing values is automatically determined (preferable a light grayscale color, but if it is indistinguishable by color blind people, a light color with a low chroma value is selected)}

\item{types}{character vector of the same length as x (or length 1), which determines the type of palette \code{"cat"}, \code{"seq"}, or \code{"div"}.}

\item{series}{for \code{c4a_palettes_add}, a character vector of the same length as x (or length 1), which determines the series. For \code{c4a_palettes_remove} a character vector of series names that should be removed (use \code{"all"} to remove all).}

\item{nmin, nmax, ndef}{minimum / maximum / default number of colors for the palette. By default: \code{nmin = 1}, for \code{"cat"} \code{nmax} and \code{ndef} the number of supplied colors. For the other types, \code{nmax} is \code{Inf}. \code{ndef} is 7 for \code{"seq"}, 9. For diverging palettes, these numbers refer to the number of columns. (See \code{mmin}, \code{mmax}, \code{mdef} for the rows)}

\item{mmin, mmax, mdef}{minimum / maximum / default number of rows for bivariate palettes.}

\item{format.palette.name}{should palette names be formatted to lowercase/underscore format?}

\item{remove.blacks, take.gray.for.NA, remove.other.grays}{These arguments determine the processing of grayscale colors for categorical \code{"cat"} palettes: if \code{remove.blacks} and there are (near) blacks, these are removed first. Next, if \code{take.gray.for.NA}, \code{xNA} is \code{NA}, and a palette contains at least one grayscale color (which can also be white), this is used as color for missing values. In case there are more than one grayscale color, the lightest is taken. \code{remove.other.grays} determines what happens with the other grays.}

\item{light.to.dark}{should sequential \code{"seq"} palettes be automatically ordered from light to dark?}

\item{remove.names}{should individual color names be removed?}

\item{biv.method}{method to a create bivariate palette. Options are \code{"byrow"} means that the colors are wrapped row-wise to a color matrix where the number of rows and columns is automatically determined, \code{"byrowX"} the same but with X (integer between 2 and 9) columns, \code{"bycol"} and \verb{"bycolX} similar but wrapped column-wise. \code{"div2seqseq"} and \verb{"div2catseq} means that colors are extracted from a divering palette. The former translates colors into a matrix with the neutral color in the diagonal, while the latter places the neutral color in the middle column. \code{"seq2uncseq"}}

\item{space}{color space in which interpolated colors are determined. Options: \code{"rgb"} (RGB) and \code{"Lab"} (CIE Lab).}

\item{range_matrix_args}{list of lists, one for each palette. Each such list specifies the range of sequential and diverging palettes, in case they are not indexed. See details.}

\item{bib}{bibtex reference in the form of a \code{utils::bibentry} object.}

\item{...}{passed on to \code{c4a_palettes_add}}

\item{fullnames}{full palette names (so in the format \code{series.palette_name})}

\item{are.you.sure}{are you sure you want to remove series?}
}
\description{
Add and remove palette series. \code{c4a_palettes_add} can be used to add own palette series and  \code{c4a_palettes_remove} to remove palette series. \code{c4a_palettes_add_as_is} is the same as \code{c4a_palettes_add}, but by default without any processing. These functions require the R package \code{colorblindcheck}.
}
\details{
Indexing: for a categorical \code{"cat"} palette, an optional \code{"index"} attribute determines which colors to use for which lengths: if the palette consists of k colors, index should be a list of k, where the i-th element is an integer vector of length i with values 1,2,...,k. See \code{c4a_info("rainbow")} and  for an example.

Range: sequential and diverging palettes are usually defined for 9+ colors. The optional \code{"range_matrix"} attribute determines that range is used for less colors. It is a n x 2 matrix where row i defines the applied range of a palette of length i. For sequential palettes a range \code{c(0,1)} means that the palette is generated (via a color ramp) between the two outermost colors. For diverging palettes, a range \code{c(x, y)} means that both sides of the palette are generated (via a color ramp) from \code{x}, which is the distance to the center color, to \code{y} which represents both outermost colors.

The range is automatically set for sequential and diverging palettes that have no \code{"index"} or \code{"range_matrix"} attribute via the parameter \code{range_matrix_args}, which is a list per palette. The arguments for a sequential palette are: \code{nmin} the minimum number of colors for which the range is reduced, \code{nmax}, the number of colors for which the range is set to \code{c(0,1)}, \code{slope_min} and \code{slope_max} determine the slopes of range reduction from a palette of length \code{nmax} to \code{nmin}, and \code{space} sets the color space for which the color ramp is applied (\code{"rgb"} or \code{"Lab"}). The arguments for a diverging palette are the same, but only one \code{slope} is used (namely for the outermost colors).

It may take some time to process, especially large categorical palettes, because of calculations of the color blind checks.
}
\examples{
#######################################
# Palettes from Statistics Netherlands
#######################################

cbsnl_cols = list(
	charts = c("#00a1cd", "#0058b8", "#afcb05", "#53a31d", "#d9328a", "#7d4791",
			   "#f39200", "#c90c0f", "#0581a2", "#163a72", "#899d0c", "#488225",
			   "#af0e80", "#56217a", "#da5914", "#9c1006"),
	map_blue = structure(c("#e1f4fd", "#c0e7ff", "#77cbe5", "#3d95d4", "#2256a0", "#143564",
				  "#09183c"), index = list('5' = 2:6, '6' = 2:7, '7' = 1:7)),
	map_green = structure(c("#f1f6de", "#edf0c7", "#c9de85", "#85bc22", "#348a3a", "#0f5f34",
				   "#114625"), index = list('5' = 2:6, '6' = 2:7, '7' = 1:7)),
	map_red = structure(c("#fedfc7", "#ffc597", "#f89e6b", "#e74d15", "#c01f26", "#82001e",
				 "#5b0708"), index = list('5' = 2:6, '6' = 2:7, '7' = 1:7)),
	map_purple = structure(c("#fbe2ed", "#f8c1d9", "#e38cbf", "#be3e8d", "#8b176f", "#490045",
					"#2d002c"), index = list('5' = 2:6, '6' = 2:7, '7' = 1:7)),
	map_purple_green = c("#490045", "#be3e8d", "#f8c1d9", "#e5e5e5", "#edf0c7", "#85bc22",
						 "#0f5f34"),
	map_red_blue = c("#82001e", "#e74d15", "#ffc597", "#e5e5e5", "#c0e7ff", "#3d95d4",
					 "#143564")
)

cbsnl_types = c("cat", rep("seq", 4), "div", "div")
nmin = c(NA, 5, 5, 5, 5, 7, 7)
nmax = c(NA, 7, 7, 7, 7, 7, 7)

cbsnl_bib = utils::bibentry(bibtype = "Misc",
					  author = person("Statistics Netherlands"),
					  title = "CBS - Statistics Netherlands",
					  url = "https://www.cbs.nl/en-gb",
					  year = 2022)


c4a_palettes_add(cbsnl_cols, xNA = "grey88",
			   nmin = nmin, nmax = nmax, ndef = nmin,
			   types = cbsnl_types, series = "cbsnl", bib = cbsnl_bib)

\dontrun{
c4a_gui(series = "cbsnl")
}

# Palettes proposed by
# M.A. Petroff (2021), Accessible Color Sequences for Data Visualization,
# https://arxiv.org/pdf/2107.02270.pdf
rgb2hex = function(x) rgb(x[1], x[2], x[3], maxColorValue = 255)

petroff = local({
	rgb6 = list(c(87,144,252),
		 c(248,156,32),
		 c(228,37,54),
		 c(150,74,139),
		 c(156,156,161),
		 c(122,33,221))

	rgb8 = list(c(24,69,251),
		 c(255,94,2),
		 c(201,31,22),
		 c(200,73,169),
		 c(173,173,125),
		 c(134,200,221),
		 c(87,141,255),
		 c(101,99,100))

	rgb10 = list(c(63,144,218),
		 c(255,169,14),
		 c(189,31,1),
		 c(148,164,162),
		 c(131,45,182),
		 c(169,107,89),
		 c(231,99,0),
		 c(185,172,112),
		 c(113,117,129),
		 c(146,218,221))


	list(petroff6 = sapply(rgb6, rgb2hex),
		 petroff8 = sapply(rgb8, rgb2hex),
		 petroff10 = sapply(rgb10, rgb2hex))
})
petroff2 = list(petroff9 = petroff$petroff10[-9])

c4a_palettes_add_as_is(petroff, xNA = NA, types = "cat", series = "petroff")
c4a_palettes_add_as_is(petroff2, xNA = petroff$petroff10[9],
					 types = "cat", series = "petroff")
}
