% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_show.R
\name{c4a_show}
\alias{c4a_show}
\title{Show cols4all palettes}
\usage{
c4a_show(
  n = NULL,
  type = c("cat", "seq", "div", "biv"),
  columns = 12,
  cvd.sim = c("none", "deutan", "protan", "tritan"),
  devel = TRUE
)
}
\arguments{
\item{n}{number of colors}

\item{columns}{number of columns in case `n` is not specified. Otherwise `columns` is set to `n`}

\item{cvd.sim}{color vision deficiency simulation: one of `"none"`, `"deutan"`, `"protan"`, `"tritan"`}

\item{x}{named list of color palettes}
}
\description{
Show cols4all palettes
}
