% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_sysdata_import.R
\name{c4a_sysdata_import}
\alias{c4a_sysdata_import}
\alias{c4a_sysdata_export}
\title{Import and export system data}
\usage{
c4a_sysdata_import(x)

c4a_sysdata_export()
}
\arguments{
\item{x}{system data}
}
\value{
`c4a_sysdata_export` returns the system data
}
\description{
Import and export system data. `c4a_sysdata_import` will import system data and overwrite the current system data. System data includes palette data and pre-calculated color-blind-friendliness scores. `c4a_sysdata_export` will export the current system data. These functions are for advanced use; the functions \code{\link{c4a_series_add}} and \code{\link{c4a_series_remove}} are user-friendly interfaces to change the loaded palettes.
}
\examples{
x = c4a_sysdata_export()
c4a_sysdata_import(x)
y = c4a_sysdata_export()
identical(x, y)
}
