% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pals.R
\name{check_cat_pal}
\alias{check_cat_pal}
\title{Check categorical palette}
\usage{
check_cat_pal(p)
}
\arguments{
\item{p}{sequential palette}
}
\value{
vector of three quality indices
}
\description{
Check categorical palette. It computes one quality indicator: the \code{min_dist}, the minimal distance between any two colors. This is computed for all three color vision deficiency types: the worst (i.e. lowest) score is returned.
}
\details{
and \code{max_step} is low, although the former is much more important.
}
