% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pals.R
\name{check_cyc_pal}
\alias{check_cyc_pal}
\title{Check cyclic palette}
\usage{
check_cyc_pal(p)
}
\arguments{
\item{p}{cyclic palette}
}
\value{
vector of three quality indices
}
\description{
Check cyclic palette. Same as \code{check_seq_pal}, but also the difference between the first and last color is considered as step
}
