% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pals.R
\name{check_div_pal}
\alias{check_div_pal}
\title{Check diverging palette}
\usage{
check_div_pal(p)
}
\arguments{
\item{p}{diverging palette with odd number of colors (so with a middle color)}
}
\value{
vector of three quality indices
}
\description{
Check diverging palette. It computes two quality indices. \code{inter_wing_dist} minimal distance between from one wing (any color) to the other wing (any color); let a step be the distance from one color to a neighboring color, then: \code{min_step} is the minimal step. These two quality indices are computed for all three color vision deficiency types: per quality indicator, the worst score is returned.
}
