% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pals.R
\name{check_seq_pal}
\alias{check_seq_pal}
\title{Check sequential palette}
\usage{
check_seq_pal(p)
}
\arguments{
\item{p}{sequential palette}
}
\value{
vector of three quality indices
}
\description{
Check sequential palette. It computes two quality indices. \code{min_step} and \code{max_step} are the minimum and maximum step respectively, where a step is the distance from one color to a neighboring color. \code{min_step} is the leading indicator: the higher, the better the palette. From palettes with equal \code{min_step}, those with the lowest \code{max_step} can be considered as better, because the steps are more uniform. These two quality indices are computed for all three color vision deficiency types: per quality indicator, the worst score is returned.
}
\details{
and \code{max_step} is low, although the former is much more important.
}
