% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{package}
\name{cols4all-package}
\alias{cols4all-package}
\alias{cols4all}
\title{cols4all overview}
\description{
Color palettes for all people, including those with color vision deficiency. Palettes, organized by source and type, are scored on several properties such as color-blind-friendliness and harmony. Furthermore each palette is assigned a distinct color for missing values. Own palettes can be loaded as well. ggplot2 scales are included.
}
\details{
This page provides a brief overview of all package functions.
}
\section{Main functions}{

\tabular{ll}{
\code{\link{c4a_gui}}\tab GUI (shiny app) to analyse the palettes \cr
\code{\link{c4a}}\tab Get the colors of a palette \cr
}
}

\section{Palette names and properties}{

\tabular{ll}{
\code{\link{c4a_palettes}}\tab Get available palette names \cr
\code{\link{c4a_series}}\tab Get available series \cr
\code{\link{c4a_meta}}\tab Get meta information (such as type and maximum number of colors) \cr
\code{\link{c4a_ls}}\tab Environment via which palette names can be browsed with auto-completion (using \code{$}) \cr
}
}

\section{Importing and exporting palettes}{

\tabular{ll}{
\code{\link{c4a_series_add}}\tab Add color palettes \cr
\code{\link{c4a_series_remove}}\tab Remove color palettes \cr
\code{\link{c4a_sysdata_import}}\tab Import system data \cr
\code{\link{c4a_sysdata_export}}\tab Export system data \cr
}
}

\section{References}{


\strong{misc}
\itemize{
\item \code{r3} and \code{r4} base R colors, from the \code{grDevices} package
\item \code{ggplot2} from the \code{ggplot2} package (also included in \code{grDevices})
\item \code{okabe} by Okabe-Ito (2002, revised 2008), Color Universal Design (CUD) - How to make figures and presentations that are friendly to Colorblind people -, \url{https://jfly.uni-koeln.de/color/} (also included in \code{grDevices})
\item \code{alphabet} from the \code{Polychroma} package (also included in \code{grDevices})
\item \code{alphabet2} by P.Green-Armytage (2010): A Colour Alphabet and the Limits of Colour Coding. Colour: Design & Creativity (5) (2010): 10, 1-23. www.aic-color.org/journal/v5/jaic_v5_06.pdf
\item \code{polychrome36} from the \code{Polychrome} package (also included in \code{grDevices})
\item \code{kelly} by K. Kelly (1965): Twenty-two colors of maximum contrast. Color Eng., 3(6), 1965. http://www.iscc.org/pdf/PC54_1724_001.pdf (obtained from the \code{pals} package)
\item \code{watlington} by John Watlington. \url{http://alumni.media.mit.edu/~wad/color/palette.html} (obtained from the \code{pals} package)
\item \code{cols25} by K. Wright \url{https://stackoverflow.com/a/9568659/1393348} (obtained from the \code{pals} package)
\item \code{glasbey} by Chris Glasbey, Gerie van der Heijden, Vivian F. K. Toh, Alision Gray (2007). Colour Displays for Categorical Images. Color Research and Application, 32, 304-309. (from the \code{pals} package)
}

\strong{hcl}

From the \code{colorspace} package, citation:

Zeileis A, Fisher JC, Hornik K, Ihaka R, McWhite CD, Murrell P, Stauffer R, Wilke CO (2020). colorspace: A Toolbox for Manipulating and Assessing Colors and Palettes. Journal of Statistical Software, 96 (1), 1-49.

\strong{brewer}

From the \code{RColorBrewer} package, by Cynthia Brewer:

Harrower, Mark, and Cynthia A. Brewer. "ColorBrewer.org: an online tool for selecting colour schemes for maps." The Cartographic Journal 40.1 (2003): 27-37.

\strong{tol}

Color schemes by Paul Tol. \url{https://personal.sron.nl/~pault/}

\strong{viridis}

Color schemes used in the Python library matplotlib:

J. D. Hunter, "Matplotlib: A 2D Graphics Environment", Computing in Science & Engineering, vol. 9, no. 3, pp. 90-95, 2007.

In R obtained via the package \code{viridisLite}

\strong{kovesi}

Color schemes by Peter Kovesi:

Peter Kovesi (2016). CET Perceptually Uniform Colour Maps. \url{https://colorcet.com/}
Peter Kovesi (2015). Good Colour Maps: How to Design Them. Arxiv. \url{https://arxiv.org/abs/1509.03700}

Obtained via the R package \code{pals} (by Kevin Wright)

\strong{wes}

Palettes from Wes Anderson movies

Obtained via the package \code{wesanderson}

\strong{carto}

Palettes by CARTO.

Obtained via the package \code{rcartocolors}

\strong{scico}

Palettes developed by Fabio Crameri

\url{http://www.fabiocrameri.ch/colourmaps.php}

Crameri, Fabio. (2018, May 8). Scientific colour maps (Version 3.0.1). Zenodo.
Crameri, Fabio. (2018). Geodynamic diagnostics, scientific visualisation and StagLab 3.0. Geosci. Model Dev. Discuss.

Obtained via the R package \code{scico}

\strong{tableau}

Palettes by Tableau Software

Obtained via the R package \code{ggthemes}

\strong{seaborn}

Palettes contained in the Python library \code{seaborn} by Michael Waskom. \url{https://seaborn.pydata.org/tutorial/color_palettes.html}
}

\seealso{
\url{https://github.com/mtennekes/cols4all}
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\concept{color}
\concept{visualization}
