% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pals.R
\name{dist_to_col}
\alias{dist_to_col}
\title{Calculate the minimum distance a palette (any color in this palette) and a color.}
\usage{
dist_to_col(pal, col)
}
\description{
Calculate the minimum distance a palette (any color in this palette) and a color.
}
