% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_color_c4a_cat}
\alias{scale_color_c4a_cat}
\title{col4all scales for ggplot2}
\usage{
scale_color_c4a_cat(
  palette = NULL,
  reverse = NULL,
  order = NULL,
  aesthetics = "colour",
  ...
)
}
\arguments{
\item{palette, reverse, order}{See \code{\link{c4a}}.}

\item{aesthetics}{The ggplot2 aesthetics to which this scale should be applied.}

\item{...}{common discrete scale parameters: `name`, `breaks`, `labels`, `na.value`, `limits` and `guide`. See \code{\link[ggplot:discrete_scale]{discrete_scale}} for more details.}
}
\description{
col4all scales for ggplot2
}
\examples{
library("ggplot2")

data("diamonds")

# find suitable palette
c4a_show(n = nlevels(diamonds$color), sort = "rank")

# select "carto.safe"
ggplot(diamonds[diamonds$price >= 15000, ], aes(x = carat, y = price, color = color)) +
	geom_point(size = 2) +
	scale_color_c4a_cat("carto.safe") +
	theme_light()
}
