% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sel_cat}
\alias{sel_cat}
\title{Select categorical palettes with specific length}
\usage{
sel_cat(x, n = NULL)
}
\arguments{
\item{x}{named list of palettes. Each list item is either a color vector or a list of color vectors for specific lengths (names should correspond to those lengths.}

\item{n}{desired number colors. When not specific, the maximum is taken.}
}
\description{
Select categorical palettes with specific length
}
