% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a.R
\docType{data}
\name{c4a_defaults}
\alias{c4a_defaults}
\alias{c4a}
\alias{c4a_na}
\title{Get a cols4all color palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
c4a_defaults

c4a(
  palette = NULL,
  n = NULL,
  type = c("cat", "seq", "div"),
  reverse = FALSE,
  order = NULL,
  contrast = NULL
)

c4a_na(palette = NULL, type = c("cat", "seq", "div"))
}
\arguments{
\item{palette}{name of the palette. See \code{\link{c4a_palettes}} for options. If omitted, the default palette is provided `c4a_defaults`}

\item{n}{number of colors. If omitted then: for type `"cat"` the maximum number of colors is returned, and for types `"seq"` and `"div"`, 9 colors.}

\item{type}{type of color palette, in case `palette` is not specified: one of `"cat"` (categorical/qualitative palette), `"seq"` (sequential palette) and `"div"` (diverging palette).}

\item{reverse}{should the palette be reversed?}

\item{order}{order of colors. Only applicable for `"cat"` palettes}

\item{contrast}{vector of two numbers that determine the range that is used for sequential and diverging palettes. Both numbers should be between 0 and 1. The first number determines where the palette begins, and the second number where it ends. For sequential palettes, 0 means the leftmost (normally lightest) color, and 1 the rightmost (often darkest) color. For diverging palettes, 0 means the middle color, and 1 both extremes. If only one number is provided, this number is interpreted as the endpoint (with 0 taken as the start). By default, it is set automatically, based on `n`. See `c4a_gui`, or the internal functions `cols4all::default_contrast_seq` and `cols4all::default_contrast_div` to see what the automatic values are.}
}
\value{
a vector of colors
}
\description{
Get a cols4all color palette. The function `c4a` returns the colors of the specified palette, and the function `c4a_na` returns the color for missing value that is associated with the specified palette.
}
\examples{
c4a_palettes("div")

c4a(type = "cat")

(pal = c4a("tol.sunset", n = 7, contrast = c(0, .6)))
colorspace::specplot(pal)

c4a("set2")

c4a("hcl.set2")

c4a("hcl.set2", n = 8)
}
\keyword{datasets}
