% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a.R
\name{c4a}
\alias{c4a}
\title{Select a cols4all color palette}
\usage{
c4a(
  palette = NULL,
  n = NULL,
  type = c("cat", "seq", "div", "biv", "cyc"),
  reverse = FALSE,
  order = NULL
)
}
\arguments{
\item{palette}{name of the palette. See \code{\link{c4a_palettes}} for options}

\item{n}{}
}
\description{
Select a cols4all color palette
}
