% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a.R
\name{c4a_meta}
\alias{c4a_meta}
\title{Get meta information from a cols4all palette}
\usage{
c4a_meta(palette, no.match = c("error", "null"))
}
\arguments{
\item{palette}{name of the palette}

\item{no.match}{what happens is no match is found? Options: `"error"`: an error is thrown, `"null"`: `NULL` is returned}
}
\value{
list with the following items: name, series, fullname, type, palette (colors), na (color), nmax, and reverse. The latter is `TRUE` when there is a `"-"` prefix before the palette name.
}
\description{
Get meta information from a cols4all palette
}
