% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a_modify.R
\name{c4a_modify}
\alias{c4a_modify}
\alias{c4a_duplicate}
\title{Edit cols4all palettes (in development)}
\usage{
c4a_modify(palette, x = NULL, xNA = NULL)

c4a_duplicate(palette, name = NA)
}
\arguments{
\item{palette}{name of the palette}

\item{x}{vector of the new colors. It should either the same length, or a named vector, where the names correspond to the index numbers. E.g. \code{c("3" = "#AABBCC")} will replace the third color with the color \code{"#AABBCC"}.}

\item{xNA}{the new color for missing values.}

\item{name}{name of new palette}
}
\description{
Edit cols4all palettes. c4a_duplicate duplicates an existing cols4all palette, and c4a_modify is used to change the colors. Use c4a_data to craete palettes from scratch.
}
\examples{
c4a_duplicate("brewer.set2", "set2_mod")
c4a_modify("set2_mod", c("4" = "#EA8AB8"))
}
\seealso{
\code{\link[=c4a_data]{c4a_data()}}
}
