% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{c4a_options}
\alias{c4a_options}
\title{Set cols4all options}
\usage{
c4a_options(...)
}
\arguments{
\item{...}{Use character values to retrieve options. To set options, either use named arguments (where the names refer to the options), a list that consists of those options.}
}
\description{
Get or set global options for c4a. Works similar as the base function \code{options}
}
\details{
| ------------- |:------------- |
| defaults		|  Default palettes |
| CBF_th		|  Parameters that label a palette as color blind friendly  |
| CBU_th		| Parameters that label a palette as color blind unfriendly |
| CrangeHarm		| Maximum chroma range for which a palette is considered harmonic |
| CrangeDisH		| Minimum chroma range for which a palette is considered disharmonic  |
| LrangeHarm		| Maximum luminance range for which a palette is considered harmonic |
| LrangeDisH		| Minimum luminance range for which a palette is considered disharmonic |
| Cintense		| Chroma of colors that are considered intense |
| Cpastel		| Chroma of colors that are considered 'pastel' |
| HwidthDivRainbow		| A diverging palette is labeled as 'rainbow hue' if HwidthL or HwidthR are at least \code{HwidthDivRainbow} |
| HwidthDivSingle		| A diverging palette is labeled as 'single hue' if HwidthL and HwidthR are at most \code{HwidthDivSingle} |
| HwidthSeqRainbow | A sequential palette is labeled as 'rainbow hue' if Hwidth is at least \code{HwidthSeqRainbow} |
| HwidthSeqSingle | A sequential palette is labeled as 'single hue' if Hwidth is at most \code{HwidthSeqSingle} |
}
