% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4a.R
\name{c4a_palettes}
\alias{c4a_palettes}
\title{List all available cols4all color palettes}
\usage{
c4a_palettes(type = c("all", "cat", "seq", "div", "biv", "cyc"))
}
\description{
List all available cols4all color palettes
}
