% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{package}
\name{cols4all-package}
\alias{cols4all-package}
\alias{cols4all}
\title{cols4all overview}
\description{
cols4all stands for: color palettes for all people, including those with color vision deficiency. Popular color palettes have been organized by source and type, and scored on several properties such as color-blind-friendliness and harmony. Furthermore each palette is assigned a distinct color for missing values. Own palettes can be loaded as well. ggplot2 scales are included.
}
\details{
This page provides a brief overview of all package functions.
}
\section{Main functions}{

\tabular{ll}{
\code{\link{c4a_gui}}\tab GUI (shiny app) to analyse the palettes \cr
\code{\link{c4a}}\tab Get the colors of a palette \cr
}
}

\section{Palette names and properties}{

\tabular{ll}{
\code{\link{c4a_palettes}}\tab Get available palette names \cr
\code{\link{c4a_series}}\tab Get available series names\cr
\code{\link{c4a_info}}\tab Get palette information (such as type and maximum number of colors) \cr
\code{\link{c4a_citation}}\tab Show how to cites palettes (with bibtex code) \cr
\code{\link{.P}}\tab Environment via which palette names can be browsed with auto-completion (using \code{$}) \cr
}
}

\section{Importing and exporting palettes}{

\tabular{ll}{
\code{\link{c4a_data}}\tab Build color palette data \cr
\code{\link{c4a_load}}\tab Load color palette data \cr
\code{\link{c4a_sysdata_import}}\tab Import system data \cr
\code{\link{c4a_sysdata_export}}\tab Export system data \cr
}
}

\seealso{
\url{https://github.com/mtennekes/cols4all}
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\concept{color}
\concept{visualization}
